move popover into its own file
