+++ packages/dev/sharedUiComponents/src/fluent/primitives/searchBox.tsx
@@ -1,7 +1,6 @@
-import { makeStyles, tokens, Popover, PopoverSurface, PopoverTrigger } from "@fluentui/react-components";
-import type { OnOpenChangeData, OpenPopoverEvents } from "@fluentui/react-components";
+import { makeStyles, tokens } from "@fluentui/react-components";
 import { SearchBar } from "./searchBar";
-import type { FunctionComponent, PropsWithChildren } from "react";
+import type { FunctionComponent } from "react";
 import { useState, useEffect } from "react";
 
 type SearchBoxProps = {
@@ -20,7 +19,7 @@ const useSearchBoxStyles = makeStyles({
         boxShadow: tokens.shadow16,
         display: "grid",
         gridTemplateRows: "auto auto 1fr",
-        // overflow: "hidden", // Prevent content overflow
+        overflow: "hidden", // Prevent content overflow
     },
     title: {
         borderBottom: `${tokens.strokeWidthThin} solid ${tokens.colorNeutralStroke2}`,
@@ -127,64 +126,5 @@ export const SearchBox: FunctionComponent<SearchBoxProps> = (props) => {
                 ))}
             </div>
         </div>
-    );
-};
-
-type PositionedPopoverProps = {
-    x: number;
-    y: number;
-    visible: boolean;
-    hide: () => void;
-};
-
-/**
- * PositionedPopover component that shows a popover at specific coordinates
- * @param props - The component props
- * @returns The positioned popover component
- */
-export const PositionedPopover: FunctionComponent<PropsWithChildren<PositionedPopoverProps>> = (props) => {
-    const [open, setOpen] = useState(false);
-
-    useEffect(() => {
-        setOpen(props.visible);
-    }, [props.visible, props.x, props.y]);
-
-    const handleOpenChange = (_: OpenPopoverEvents, data: OnOpenChangeData) => {
-        setOpen(data.open);
-
-        if (!data.open) {
-            props.hide();
-        }
-    };
-
-    return (
-        <>
-            <Popover
-                open={open}
-                onOpenChange={handleOpenChange}
-                positioning={{
-                    position: "below", // Places the popover directly below the trigger element
-                    align: "center", // Centers the popover horizontally relative to the trigger element
-                    autoSize: "height-always", //Automatically adjusts the popover height to fit within the viewport
-                    fallbackPositions: ["above", "after", "before"], //If the primary position doesn't fit, automatically tries these positions in order
-                }}
-                withArrow={false} // Removes arrow that points to trigger element
-            >
-                <PopoverTrigger>
-                    {/* Use the invisible div as the trigger location*/}
-                    <div
-                        style={{
-                            position: "absolute",
-                            left: `${props.x}px`,
-                            top: `${props.y}px`,
-                            width: 1,
-                            height: 1,
-                            pointerEvents: "none", // so it's invisible to interaction
-                        }}
-                    />
-                </PopoverTrigger>
-                <PopoverSurface>{props.children}</PopoverSurface>
-            </Popover>
-        </>
     );
 };
